Slight renaming, add object node
